requirements.txt
Flask==1.1.2
pandas==1.1.5
scikit-surprise==1.1.1
